panorama_vr/requirements.txt
opencv-python
imutils
numpy
Pillow